+++ src/contact/domain/contact.h
@@ -6,16 +6,16 @@
 
 class Contact
 {
-    long id;
+    int id;
     std::string name;
     std::string email;
     std::string phone;
 
 public:
     Contact();
-    Contact(long id, std::string name, std::string email, std::string phone);
+    Contact(int id, std::string name, std::string email, std::string phone);
 
-    long getId();
+    int getId();
     std::string getName();
     std::string getEmail();
     std::string getPhone();
